Update elements to allow optional storage of text as clob or varrays
on 11g when these objects are used within static sql varchar2 strings
are limited to 4000 char.  I knew that but seldom have text values of
that size or greater.  In this case you will need to use the clob
format.  In 12c this changes to 32k.  However my experiments show that
clobs under 32k function behind the scenes as varchar2 anyhow.  As this
library is targeted for 11g at least for now I am leaving the varchar2
items set to max out at 4000.